# docs/swagger.yaml
definitions:
...
  model.Shop:
    properties:
      current_num:
        type: string
      fiery_num:
        description: 火热指数
        type: string
      hot_line:
        type: string
      id:
        type: integer
      opening_time:
        description: 营业时间
        type: string
      path:
        type: string
      picture:
        description: 商店图片
        type: string
      service_intro:
        description: 服务介绍
        type: string
      sha:
        type: string
      shop_name:
        description: 商店名称
        type: string
      should_know:
        description: 需知
        type: string
      vip_service:
        description: vip服务
        type: string
    type: object
...
  user.ScriptId:
    properties:
      script_id:
        description: 剧本的ID
        type: string
    type: object
  user.Status:
    properties:
      status:
        description: 标签
        type: string
    type: object
  user.UserInfor:
    properties:
      email:
        description: 邮箱
        type: string
      gender:
        description: 性别
        type: string
      idcard:
        description: 身份证号
        type: string
      nickname:
        description: 用户名
        type: string
      phone:
        description: 电话号码
        type: string
      realname:
        description: 真实姓名
        type: string
    type: object
host: localhost:8918
info:
  contact:
    email: 2295616516@qq.com
    name: Eternal-Faith
  description: 山乡剧变
  termsOfService: http://swagger.io/terrms
  title: uima
  version: 1.0.0
paths:
...
  /shop:
    post:
      consumes:
      - application/json
      description: '"新增店铺信息"'
      parameters:
      - description: shop
        in: body
        name: shop
        required: true
        schema:
          $ref: '#/definitions/model.Shop'
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功
        "400":
          description: 上传失败
        "401":
          description: 身份验证失败
      summary: '"店铺信息"'
      tags:
      - shop
  /shop/:shop_id:
    delete:
      consumes:
      - application/json
      description: '"删除一个店铺的信息"'
      parameters:
      - description: shop_id
        in: path
        name: shop_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
        "400":
          description: 删除失败
      summary: '"删除店铺信息"'
      tags:
      - shop
  /shop/get_all:
    get:
      consumes:
      - application/json
      description: '"获取商店详细信息"'
      parameters:
      - description: shop
        in: body
        name: shop
        required: true
        schema:
          $ref: '#/definitions/model.Shop'
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功
        "400":
          description: 上传失败
        "401":
          description: 身份验证失败
      summary: '"商店信息"'
      tags:
      - shop
  /shop/single_shop/:shop_id:
    get:
      consumes:
      - application/json
      description: '"获取单个商店详细信息"'
      parameters:
      - description: shop_id
        in: path
        name: shop_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
        "400":
          description: 获取失败
        "401":
          description: 获取参数失败
      summary: '"单个商店信息"'
      tags:
      - shop
...
  /user/cancel:
    post:
      consumes:
      - application/json
      description: '"将收藏的剧本取消"'
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 剧本Id
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/user.ScriptId'
      produces:
      - application/json
      responses:
        "200":
          description: 取消收藏成功
        "400":
          description: 发送请求失败
        "401":
          description: 身份验证失败
        "500":
          description: 服务器错误，取消收藏收藏失败
      summary: '"取消收藏剧本"'
      tags:
      - user
...
  /user/collection:
    post:
      consumes:
      - application/json
      description: '"将该剧本收藏到我的收藏中"'
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 剧本Id
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/user.ScriptId'
      produces:
      - application/json
      responses:
        "200":
          description: 收藏成功
        "401":
          description: 身份验证失败
        "500":
          description: 服务器错误，收藏失败
      summary: '"收藏剧本"'
      tags:
      - user
  /user/edit:
    post:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: user
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/user.UserInfor'
      responses:
        "200":
          description: 登陆成功
        "400":
          description: 输入格式错误
        "401":
          description: 身份认证失效
        "404":
          description: 用户不存在
        "500":
          description: 服务器错误
      summary: '"设置个人信息"'
      tags:
      - user
...
  /user/mycollection:
    get:
      consumes:
      - application/json
      description: '"获取用户的所有剧本收藏"'
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 查看成功
        "400":
          description: 发送请求失败
        "401":
          description: 身份认证失效
        "500":
          description: 服务器错误
      summary: '"我的收藏(剧本)"'
      tags:
      - user
  /user/order:
    post:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 剧本Id
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/user.ScriptId'
      responses:
        "200":
          description: 创建订单成功
        "400":
          description: 输入格式有误
        "500":
          description: 服务器错误
      summary: '"创建订单(剧本)"'
      tags:
      - user
...
  /user/searchappointment:
    post:
      consumes:
      - application/json
      description: '"通过标签来进行分类"'
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 预约状态
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/user.Status'
      produces:
      - application/json
      responses:
        "200":
          description: 搜索成功
        "400":
          description: 输入格式有误
        "404":
          description: 搜索失败
        "500":
          description: 服务器错误
      summary: '"剧本分类"'
      tags:
      - user
...
  /user/vieworder:
    get:
      consumes:
      - application/json
      description: '"查看所有订单(含时间，价格，剧本名字)"'
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
        "401":
          description: 身份验证错误
        "500":
          description: 服务器错误
      summary: '"查看订单"'
      tags:
      - user
schemes:
- http
swagger: "2.0"